requirements.txt
Flask~=2.0.1
PyJWT~=2.1.0
Werkzeug~=2.0.1
pymongo~=3.12.0
gunicorn~=20.1.0
Jinja2
MarkupSafe
click
colorama
dnspython
itsdangerous
jsonify
numpy~=1.21.2
pip
setuptools
wheel


requests~=2.26.0